# docs/swagger.yaml
definitions:
  handler.CreateOpportunityRequest:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  handler.CreateOpportunityResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpportunityResponse'
      message:
        type: string
    type: object
  handler.DeleteOpportunityResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpportunityResponse'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.ListOpportunitiesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.OpportunityResponse'
        type: array
      message:
        type: string
    type: object
  handler.ShowOpportunityResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpportunityResponse'
      message:
        type: string
    type: object
  handler.UpdateOpportunityRequest:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  handler.UpdateOpportunityResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpportunityResponse'
      message:
        type: string
    type: object
  schemas.OpportunityResponse:
    properties:
      company:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /opportunities:
    get:
      consumes:
      - application/json
      description: List all Job Opportunities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListOpportunitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Opportunities
      tags:
      - Opportunities
  /opportunity:
    delete:
      consumes:
      - application/json
      description: Delete a new Job Opportunity
      parameters:
      - description: Opportunity Identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteOpportunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Opportunity
      tags:
      - Opportunities
    get:
      consumes:
      - application/json
      description: Show a Job Opportunity
      parameters:
      - description: Opportunity Identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowOpportunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Show Opportunity
      tags:
      - Opportunities
    post:
      consumes:
      - application/json
      description: Create a new Job Opportunity
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOpportunityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateOpportunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Opportunity
      tags:
      - Opportunities
    put:
      consumes:
      - application/json
      description: Update a Job Opportunity
      parameters:
      - description: Opportunity Identification
        in: query
        name: id
        required: true
        type: string
      - description: opportunity data to update
        in: body
        name: opportunity
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOpportunityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateOpportunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Opportunity
      tags:
      - Opportunities
swagger: "2.0"